requirements.txt
altgraph==0.17.4
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.2
kiwisolver==1.4.7
numpy==2.2.0
packaging==24.2
pandas==2.2.3
pefile==2023.2.7
pillow==11.0.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32-ctypes==0.2.3
six==1.17.0
speedtest-cli==2.1.3
tabulate==0.9.0
tzdata==2024.2
